removing startup
