+++ Game/math.cpp
@@ -1,6 +1,18 @@
 #include "math.h"
 #include "intrinsics.h"
 
+// Utilities
+// -----------------------------------------------------------------------------
+
+uint32 gcd(uint32 a, uint32 b)
+{
+    if(b == 0){
+        return a;
+    } else{
+        return gcd(b, a % b);
+    }
+}
+
 // 2D vectors.
 // -----------------------------------------------------------------------------
