+++ Startup/Startup.vcxproj
@@ -151,9 +151,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="macros.h" />
     <ClInclude Include="startup.h" />
-    <ClInclude Include="types.h" />
     <ClInclude Include="utilities.h" />
   </ItemGroup>
   <ItemGroup>
