+++ Startup/startup.h
@@ -8,8 +8,6 @@
 // the platform layer only. All other definitions reside in either the game
 // layer or the platform layer exclusively.
 
-#include "macros.h"
-#include "types.h"
-#include "utilities.h"
+
 
 #endif
